[None][chore] Make sampler_common a dependency-free base layer
sampler_common previously imported sampler_strategy for a single function,
which made it look like a layer *under* strategy when it is not. Restructure
so the split follows what each piece actually needs:

sampler_common now imports nothing from the package and holds everything a
module can need without resolving a sampling strategy -- the per-request
queries (_request_get_sampling_params and friends), UtilsSamplingParams
(the plain scalar view of a request's sampling config, which is the *input*
to strategy resolution, not a product of it), the shared DEFAULT_BEAM_IDX /
DEFAULT_STEP_IDX constants and FinishReasonsList, plus the existing tensor
helpers.

sampler_strategy keeps _request_strategy, the one request-facing helper that
has to know what a Strategy is, and now reads its inputs from sampler_common.

This is what feature modules need: the in-flight penalties (#16485) and beam
search (#16620) work both hit circular imports because the per-request
helpers lived in sampler.py, and neither needs Strategy -- penalties uses
_unwrap_singleton / _get_max_beam_width / DEFAULT_BEAM_IDX and never resolves
a strategy. Those now sit in a layer any feature module can import.

top_p_decay drops to sampler_common for the request query and takes Fusions
straight from ops.vanilla (sampler_strategy only re-exported it), leaving
TopPDecayMetadata as its one genuine strategy dependency.

No behavior change.
